Working for MIYOO, W32 and PSP
Added new classes to build vcxproj
Had to add ancient assignment of vectors in RandomNames.cpp due to <C++11

Consider bumping build system to something more modern
