+++ sources/Application/Views/ModalDialogs/NewProjectDialog.cpp
@@ -1,17 +1,16 @@
 
 #include "NewProjectDialog.h"
-#include "Application/Utils/RandomName.h"
+#include "Application/Utils/RandomNames.h"
 
 static char *buttonText[BUTTONS_LENGTH] = {
-	"Regen",
-	"Ok",
-	"Cancel"
+	(char *)"Regen",
+	(char *)"Ok",
+	(char *)"Cancel"
 } ;
 
 #define DIALOG_WIDTH 20
 
-NewProjectDialog::NewProjectDialog(View &view):ModalView(view) {
-}
+NewProjectDialog::NewProjectDialog(View &view):ModalView(view) {}
 
 NewProjectDialog::~NewProjectDialog() {
 }
@@ -61,7 +60,7 @@ void NewProjectDialog::OnPlayerUpdate(PlayerEventType ,unsigned int currentTick)
 void NewProjectDialog::OnFocus() {
 	selected_ = 1;
 	currentChar_ = 0;
-	std::string randomName = getRandomName();
+	std::string randomName = RandomNames::getRandomName();
 	memset(name_,' ',MAX_NAME_LENGTH+1) ;
 	strncpy(name_, randomName.c_str(), sizeof(randomName) - 1);
 	lastChar_ = randomName.c_str()[0];
@@ -79,7 +78,7 @@ void NewProjectDialog::ProcessButtonMask(unsigned short mask,bool pressed) {
 	  // A modifier
 	  if (mask&EPBM_A) { 
 		if (mask==EPBM_A) {
-			std::string randomName = getRandomName();
+			std::string randomName = RandomNames::getRandomName();
 			switch(selected_) {
 				case 0:
 					if (name_[currentChar_]==' ') {
